synced with latest, handle unicode piped to output file
Need to improve unicode handling with stdout so that the characters still show when not logging stdout to a file. But at least it runs and I can script it to run on a schedule & log output to a file. No longer using Substack API (which has restacks) although the code is still in there. Restack counts are not available in this version.