fix bug in saving article titles with slashes
Replace forward and backward slashes in article titles with _ before trying to save file to temp. We still allow them in the temp_folder spec. Having slashes in the article title was causing it to be treated like a folder which didn't exist (and which we did not want to create)